Compute Win32 window size before displaying window
Previously, setting the default window resolution _after_ the call to
ShowWindow would cause the window to jitter on startup as it's being
resized. Compute the window size during the `WM_CREATE` message instead,
after the UI elements have been created (so their heights can be used
in the calculation) but before the window is actually displayed.